%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ./edgesToMerge.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{edgesToMerge}
\alias{edgesToMerge}


\title{ Identifies edges that should be merged to parse a NCI network}


\description{
  Identifies edges that should be merged to parse a NCI network.
}

\usage{edgesToMerge(g)}

\arguments{
  \item{g}{A \code{\link[=graph-class]{graph}} object.}
}

\value{A \code{\link[base]{list}} of edges to be merged}

\author{Laurent Jacob}

\seealso{
  \code{\link{parseNCInetwork}}()
}

