%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ./is.NCIgraph.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{is.NCIgraph}
\alias{is.NCIgraph}


\title{Assess whether a graph is a NCI graph}


\description{
 Assess whether a graph is a NCI graph.
}

\usage{is.NCIgraph(gr)}

\arguments{
  \item{gr}{A \code{\link[=graph-class]{graph}} object.}
}

\value{A \code{\link[base]{logical}}, \code{\link[base:logical]{TRUE}} if the graph is a NCI graph, \code{\link[base:logical]{FALSE}} otherwise.}

\author{Laurent Jacob}

\seealso{
  \code{\link{parseNCInetwork}}()
}

