\name{NCIgraph}
\docType{class}
\alias{NCIgraph-class}


 \title{Class NCIgraph}

 \description{
   Package:  NCIgraph \cr
   \bold{Class NCIgraph}\cr
   
   public static class \bold{NCIgraph}\cr
   extends graphNEL{Object}\cr
   
   Class extending graphNEL fro graphs build from NCI gene networks.
 }
 
 \author{Laurent Jacob}
 
 \keyword{classes}
 \keyword{documentation}
 